HookでAppThemeを取得できる形に修正
