+++ example-app/SantokuApp/package.json
@@ -53,6 +53,7 @@
   "devDependencies": {
     "@babel/core": "^7.9.0",
     "@testing-library/jest-native": "^4.0.2",
+    "@testing-library/react-hooks": "^7.0.2",
     "@testing-library/react-native": "^8.0.0",
     "@types/jest": "<26.0.0",
     "@types/react": "~16.9.35",
