+++ example-app/SantokuApp/src/components/contexts/theme/index.ts
@@ -1,0 +1,2 @@
+export {useAppTheme} from './useAppTheme';
+export {AppThemeProvider} from './AppThemeContext';
